# [WIP] Allow the use of an Arena for allocating Docs