This PR changes the way documents are created to allow them to be created with an Arena from [typed_arena](https://github.com/SimonSapin/rust-typed-arena). By using an Arena instead of `Box` pretty printing (creation of Docs + writing the Docs) can get a speedup of up to 2x (see benchmarks).

Some bug fixes and ergonomic improvements are also included, see individual commits for details.