@Marwes 2016-08-14
Already got merge access (thanks @freebroccolo) so I will merge myself when I am happy with it. Posting as a PR in any case to show the changes beforehand for anyone interested.

Benchmarks

Before (just added arena allocation)
```
test bench_io_arena         ... bench:      15,587 ns/iter (+/- 9,908)
test bench_io_box           ... bench:      25,463 ns/iter (+/- 13,068)
test bench_large_io_arena   ... bench:   1,228,050 ns/iter (+/- 779,835)
test bench_large_io_box     ... bench:   1,687,678 ns/iter (+/- 395,438)
test bench_large_sink_arena ... bench:     869,787 ns/iter (+/- 265,636)
test bench_large_sink_box   ... bench:   1,581,701 ns/iter (+/- 955,118)
test bench_large_vec_arena  ... bench:   1,175,236 ns/iter (+/- 474,348)
test bench_large_vec_box    ... bench:   1,552,222 ns/iter (+/- 320,769)
test bench_sink_arena       ... bench:      12,467 ns/iter (+/- 2,049)
test bench_sink_box         ... bench:      20,538 ns/iter (+/- 4,326)
test bench_vec_arena        ... bench:      12,530 ns/iter (+/- 2,745)
test bench_vec_box          ... bench:      19,976 ns/iter (+/- 4,898)

After
```
test bench_io_arena         ... bench:       9,277 ns/iter (+/- 4,266)
test bench_io_box           ... bench:      20,075 ns/iter (+/- 13,234)
test bench_large_io_arena   ... bench:     861,352 ns/iter (+/- 329,106)
test bench_large_io_box     ... bench:   1,336,097 ns/iter (+/- 693,094)
test bench_large_sink_arena ... bench:     448,636 ns/iter (+/- 78,160)
test bench_large_sink_box   ... bench:     800,791 ns/iter (+/- 240,076)
test bench_large_vec_arena  ... bench:     667,919 ns/iter (+/- 158,713)
test bench_large_vec_box    ... bench:   1,082,055 ns/iter (+/- 259,704)
test bench_sink_arena       ... bench:       7,737 ns/iter (+/- 1,122)
test bench_sink_box         ... bench:      15,297 ns/iter (+/- 2,271)
test bench_vec_arena        ... bench:       7,712 ns/iter (+/- 2,059)
test bench_vec_box          ... bench:      16,494 ns/iter (+/- 4,184)
```