@Marwes 2016-08-16
Pretty happy with the current additions so going to merge and push a new (breaking) version to crates.io. Probably there should be no changes necessary for users of earlier versions but given the extent of the changes (and my relative unfamiliarity with the library) I'd rather be safe than sorry.